lab3/requirements.txt
ray
ray[air]
matplotlib
torch
torchvision
torchaudio
python-mnist